Refactor set config into keras_utils.
