+++ official/resnet/keras/keras_imagenet_main.py
@@ -93,23 +93,11 @@ def run(flags_obj):
   Returns:
     Dictionary of training and eval stats.
   """
-  # TODO(tobyboyd): Remove eager flag when tf 1.0 testing ends.
-  # Eager is default in tf 2.0 and should not be toggled
-  if keras_common.is_v2_0():
-    keras_utils.set_config_v2(
-        enable_xla=flags_obj.enable_xla,
-        enable_grappler_layout_optimizer=
-        flags_obj.enable_grappler_layout_optimizer)
-  else:
-    config = keras_utils.get_config_proto_v1(
-        enable_xla=flags_obj.enable_xla,
-        enable_grappler_layout_optimizer=
-        flags_obj.enable_grappler_layout_optimizer)
-    if flags_obj.enable_eager:
-      tf.compat.v1.enable_eager_execution(config=config)
-    else:
-      sess = tf.Session(config=config)
-      tf.keras.backend.set_session(sess)
+  keras_utils.set_session_config(
+      enable_eager=flags_obj.enable_eager,
+      enable_xla=flags_obj.enable_xla,
+      enable_grappler_layout_optimizer=
+      flags_obj.enable_grappler_layout_optimizer)
 
   # Execute flag override logic for better model performance
   if flags_obj.tf_gpu_thread_mode:
@@ -268,6 +256,6 @@ def main(_):
 
 if __name__ == '__main__':
   tf.compat.v1.logging.set_verbosity(tf.compat.v1.logging.INFO)
-  imagenet_main.define_imagenet_flags(dynamic_loss_scale=True)
+  imagenet_main.define_imagenet_flags(dynamic_loss_scale=True, enable_xla=True)
   keras_common.define_keras_flags()
   absl_app.run(main)
